backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Auth
python-jose[cryptography]==3.3.0
python-multipart==0.0.17

# OpenAI & LangChain
openai==1.55.3
langchain==0.3.9
langchain-openai==0.2.8
langchain-community==0.3.9

# Document handling
pypdf==5.1.0
python-dotenv==1.0.1
